eth/filters: move more code to filter_system.go
